pyproject.toml
[project]
name = "social_gihub"
version = "0.1.0"
description = ""
authors = [
    {name = "Forgot-0",email = "mr.aisamutdin@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "aiosmtplib (>=5.1.0,<6.0.0)",
    "alembic (>=1.18.4,<2.0.0)",
    "taskiq[redis] (>=0.12.1,<0.13.0)",
    "taskiq-redis (>=1.2.2,<2.0.0)",
    "uvicorn (>=0.42.0,<0.43.0)",
    "orjson (>=3.11.7,<4.0.0)",
    "dishka (>=1.9.1,<2.0.0)",
    "fastapi-limiter (==0.1.6)",
    "pydantic-settings (>=2.13.1,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "aiocache (>=0.12.3,<0.13.0)",
    "jinja2 (>=3.1.6,<4.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pydantic[email] (>=2.12.5,<3.0.0)",
    "fastapi[standard] (>=0.135.2,<0.136.0)",
    "structlog (>=25.5.0,<26.0.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "aiokafka (>=0.12.0,<0.13.0)",
    "tenacity (>=9.1.2,<10.0.0)",
    "user-agents (>=2.2.0,<3.0.0)",
    "minio (>=7.2.18,<8.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "prometheus-client (>=0.23.1,<0.24.0)",
    "prometheus-fastapi-instrumentator (>=7.1.0,<8.0.0)",
    "pyvips[binary] (>=3.1.1,<4.0.0)",
    "python-magic (>=0.4.27,<0.5.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "aiojobs (>=1.4.0,<2.0.0)",
    "websockets (>=16.0,<17.0)",
]

[dependency-groups]
test = [
    "pytest==8.3.*",
    "pytest-asyncio>=0.26.0",
    "pytest-order>=1.3.0",
    "pytest-xdist>=3.6.1",
    "testcontainers>=4.10.0"
]
lint = [
    "mypy==1.15.*",
    "ruff==0.9.*",
    "pylint==3.3.*"
]
dev = [
    "pre-commit==4.5.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
